Vulkan details
The Vulkan management part of this engine is under development (and since everything takes longer with Vulkan ...), so without adding additional source code it would be impossible to make vulkan work in depth. (more or less further than a simple rectangle), do not hesitate to collaborate, it would be with great pleasure!